frontend/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
docutils==0.19
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
pkg_resources==0.0.0
Pygments==2.13.0
requests==2.28.1
urllib3==1.26.13
